Add Generated Keys To Created Objects

+++ src/main/java/models/ExampleObject.java
@@ -12,7 +12,7 @@ import java.util.List;
 public class ExampleObject extends BaseObject {
 
 	public final static List<Attribute> Attributes = asList(
-			new Attribute("Id", int.class, "id", JDBCType.INTEGER, true),
+			new Attribute("Id", Long.class, "id", JDBCType.INTEGER, true),
 			new Attribute("Name", String.class, "name", JDBCType.VARCHAR, false)
 	);
 	public final static String TableName = "example_table";
@@ -27,6 +27,6 @@ public class ExampleObject extends BaseObject {
 		return TableName;
 	}
 
-	public int Id;
+	public Long Id;
 	public String Name;
 }
